Split Core and Project
