+++ middleware/middleware/__init__.py
@@ -1,2 +1,2 @@
 from .server import *
-from .api import endpoints
+from .api import core_api, project_api
